% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_setup.R
\name{dispersal_matrix}
\alias{dispersal_matrix}
\title{Generate Dispersal Matrix}
\usage{
dispersal_matrix(
  landscape,
  torus = TRUE,
  disp_mat,
  kernel_exp = 0.1,
  plot = TRUE
)
}
\arguments{
\item{landscape}{landscape generated by landscape_generate()}

\item{torus}{whether to model the landscape as a torus}

\item{disp_mat}{optional matrix with each column specifying the probability that an individual disperses to each other patch (row)}

\item{kernel_exp}{the exponential rate at which dispersal decreases as a function of the distance between patches}

\item{plot}{option to show plot of environmental variation}
}
\value{
matrix with dispersal probabilities
}
\description{
Generates dispersal matrix for metacommunity simulations
}
\examples{
dispersal_matrix(landscape_generate())

}
\author{
Patrick L. Thompson, \email{patrick.thompson@zoology.ubc.ca}
}
