% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_setup.R
\name{env_generate}
\alias{env_generate}
\title{Generate Environment}
\usage{
env_generate(landscape, env.df, env1Scale = 500, timesteps = 1000, plot = TRUE)
}
\arguments{
\item{landscape}{landscape generated by landscape_generate()}

\item{env.df}{optional dataframe with environmental conditions with columns: env1, patch, time}

\item{env1Scale}{scale of environmental autocorrelation between 0 and 1000}

\item{timesteps}{number of timesteps to simulate}

\item{plot}{option to show plot of environmental variation}
}
\value{
dataframe with x and y coordinates, time, and environmental conditions
}
\description{
Generates density independent environmental conditions for metacommunity simulations
}
\examples{
env_generate(landscape_generate())

}
\author{
Patrick L. Thompson, \email{patrick.thompson@zoology.ubc.ca}
}
