% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_setup.R
\name{env_traits}
\alias{env_traits}
\title{Generate Species Env. Traits}
\usage{
env_traits(
  species,
  max_r = 5,
  min_env = 0,
  max_env = 1,
  env_niche_breadth = 0.5,
  optima,
  plot = TRUE,
  optima_spacing = "random"
)
}
\arguments{
\item{species}{number of species to simulate}

\item{max_r}{intrinsic growth rate in optimal environment, can be single value or vector of length species}

\item{min_env}{minium environmental optima}

\item{max_env}{minium environmental optima}

\item{env_niche_breadth}{standard deviation of environmental niche breadth, can be single value or vector of length species}

\item{optima}{optional values of environmental optima, should be a vector of length species}

\item{plot}{option to show plot of environmental variation}

\item{optima_spacing}{"even" or "random" to specify how optima should be distributed}
}
\value{
dataframe an optima and niche breadth for each species
}
\description{
Generates species specific traits for density independent environmental responses
}
\examples{
env_traits(species = 10)

}
\author{
Patrick L. Thompson, \email{patrick.thompson@zoology.ubc.ca}
}
