% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_setup.R
\name{landscape_generate}
\alias{landscape_generate}
\title{Generate landscape}
\usage{
landscape_generate(patches = 100, xy, plot = TRUE)
}
\arguments{
\item{patches}{number of patches to include in landscape}

\item{xy}{optional dataframe with x and y columns for patch coordinates}

\item{plot}{option to show plot of landscape}
}
\value{
landscape with x and y coordinates
}
\description{
Generates a landscape for metacommunity simulations
}
\examples{
landscape_generate()

}
\author{
Patrick L. Thompson, \email{patrick.thompson@zoology.ubc.ca}
}
