% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_setup.R
\name{species_int_mat}
\alias{species_int_mat}
\title{Generate Species Interaction Matrix}
\usage{
species_int_mat(
  species,
  intra = 1,
  min_inter = 0,
  max_inter = 1.5,
  int_matrix,
  comp_scaler = 0.05,
  plot = TRUE
)
}
\arguments{
\item{species}{number of species to simulate}

\item{intra}{intraspecific competition coefficient, single value or vector of length species}

\item{min_inter}{min interspecific comp. coefficient}

\item{max_inter}{max interspecific comp. coefficient}

\item{int_matrix}{option to supply externally generated competition matrix}

\item{comp_scaler}{value to multiply all competition coefficients by}

\item{plot}{option to show plot of competition coefficients}
}
\value{
species interaction matrix
}
\description{
Generates density dependent matrix of per capita competition
}
\examples{
env_traits(species = 10)

}
\author{
Patrick L. Thompson, \email{patrick.thompson@zoology.ubc.ca}
}
